Additional refactoring  - namespace adjustments  - extracted internal classes
