+++ StylableWinFormsControls/StylableWinFormsControls/Controls/StylableDataGridView.cs
@@ -1,88 +1,87 @@
-﻿namespace AssortedDevelopment.StylableWinFormsControls
+﻿namespace StylableWinFormsControls;
+
+public class StylableDataGridView : DataGridView
 {
-    public class StylableDataGridView : DataGridView
+    private Form _parentForm;
+
+    public new bool DoubleBuffered
     {
-        private Form _parentForm;
+        get => base.DoubleBuffered;
+        set => base.DoubleBuffered = value;
+    }
 
-        public new bool DoubleBuffered
+    // Override default scrollbar to make styling it possible.
+    public new ScrollBars ScrollBars { get; set; }
+
+    public bool EnableFirstColumnGrouping { get; set; } = true;
+
+    protected Form ParentForm => _parentForm ??= FindForm();
+
+    public StylableDataGridView()
+    {
+        DoubleBuffered = true;
+    }
+
+    protected override void OnCellPainting(DataGridViewCellPaintingEventArgs args)
+    {
+        base.OnCellPainting(args);
+
+        if (!EnableFirstColumnGrouping)
         {
-            get => base.DoubleBuffered;
-            set => base.DoubleBuffered = value;
+            return;
         }
 
-        // Override default scrollbar to make styling it possible.
-        public new ScrollBars ScrollBars { get; set; }
+        // Ignore column and row headers, first row, not sorted columns and not-frozen columns 
+        if (args.RowIndex < 1 || args.RowIndex >= Rows.Count - 1 || args.ColumnIndex < 0 || args.ColumnIndex != SortedColumn?.Index || !Columns[args.ColumnIndex].Frozen)
+            return;
 
-        public bool EnableFirstColumnGrouping { get; set; } = true;
+        args.AdvancedBorderStyle.Top =
+            DataGridViewAdvancedCellBorderStyle.None;
 
-        protected Form ParentForm => _parentForm ??= FindForm();
+        args.AdvancedBorderStyle.Bottom =
+            IsRepeatedCellValue(args.RowIndex + 1, args.ColumnIndex)
+                ? DataGridViewAdvancedCellBorderStyle.None
+                : AdvancedCellBorderStyle.Bottom;
+    }
 
-        public StylableDataGridView()
+
+    protected override void OnCellFormatting(DataGridViewCellFormattingEventArgs args)
+    {
+        // Call home to base
+        base.OnCellFormatting(args);
+
+        // First row always gets displayed
+        if (args.RowIndex == 0 || args.ColumnIndex != SortedColumn?.Index || !Columns[args.ColumnIndex].Frozen)
+            return;
+
+        if (IsRepeatedCellValue(args.RowIndex, args.ColumnIndex))
         {
-            DoubleBuffered = true;
+            args.Value = string.Empty;
+
+            args.CellStyle.Font = new Font(args.CellStyle.Font, FontStyle.Bold);
+            args.FormattingApplied = true;
         }
+    }
 
-        protected override void OnCellPainting(DataGridViewCellPaintingEventArgs args)
+    private bool IsRepeatedCellValue(int rowIndex, int colIndex)
+    {
+        if (!EnableFirstColumnGrouping)
         {
-            base.OnCellPainting(args);
-
-            if (!EnableFirstColumnGrouping)
-            {
-                return;
-            }
-
-            // Ignore column and row headers, first row, not sorted columns and not-frozen columns 
-            if (args.RowIndex < 1 || args.RowIndex >= Rows.Count - 1 || args.ColumnIndex < 0 || args.ColumnIndex != SortedColumn?.Index || !Columns[args.ColumnIndex].Frozen)
-                return;
-
-            args.AdvancedBorderStyle.Top =
-              DataGridViewAdvancedCellBorderStyle.None;
-
-            args.AdvancedBorderStyle.Bottom =
-                IsRepeatedCellValue(args.RowIndex + 1, args.ColumnIndex)
-                    ? DataGridViewAdvancedCellBorderStyle.None
-                    : AdvancedCellBorderStyle.Bottom;
-        }
-
-
-        protected override void OnCellFormatting(DataGridViewCellFormattingEventArgs args)
-        {
-            // Call home to base
-            base.OnCellFormatting(args);
-
-            // First row always gets displayed
-            if (args.RowIndex == 0 || args.ColumnIndex != SortedColumn?.Index || !Columns[args.ColumnIndex].Frozen)
-                return;
-
-            if (IsRepeatedCellValue(args.RowIndex, args.ColumnIndex))
-            {
-                args.Value = string.Empty;
-
-                args.CellStyle.Font = new Font(args.CellStyle.Font, FontStyle.Bold);
-                args.FormattingApplied = true;
-            }
-        }
-
-        private bool IsRepeatedCellValue(int rowIndex, int colIndex)
-        {
-            if (!EnableFirstColumnGrouping)
-            {
-                return false;
-            }
-
-            DataGridViewCell currentCell =
-                Rows[rowIndex].Cells[colIndex];
-            DataGridViewCell prevCell =
-                Rows[rowIndex - 1].Cells[colIndex];
-
-            if (currentCell.Value == prevCell.Value ||
-                currentCell.Value != null && prevCell.Value != null &&
-                currentCell.Value.ToString() == prevCell.Value.ToString())
-            {
-                return true;
-            }
-
             return false;
         }
+
+        DataGridViewCell currentCell =
+            Rows[rowIndex].Cells[colIndex];
+        DataGridViewCell prevCell =
+            Rows[rowIndex - 1].Cells[colIndex];
+
+        if (currentCell.Value == prevCell.Value ||
+            currentCell.Value != null && prevCell.Value != null &&
+            currentCell.Value.ToString() == prevCell.Value.ToString())
+        {
+            return true;
+        }
+
+        return false;
     }
 }
