+++ StylableWinFormsControls/StylableWinFormsControls/Controls/StylableTextBox.cs
@@ -1,151 +1,150 @@
 ﻿using System.ComponentModel;
 using Timer = System.Windows.Forms.Timer;
 
-namespace AssortedDevelopment.StylableWinFormsControls
+namespace StylableWinFormsControls;
+
+public class StylableTextBox : TextBox
 {
-    public class StylableTextBox : TextBox
+    private Timer m_delayedTextChangedTimer;
+    private string _hint;
+
+    public event EventHandler DelayedTextChanged;
+    private EventHandler _hintActiveChanged;
+    /// <summary>
+    /// will be triggered when IsHintActive changes.
+    /// this event will detect when a callback is already registered and will not register again
+    /// </summary>
+    public event EventHandler HintActiveChanged
     {
-        private Timer m_delayedTextChangedTimer;
-        private string _hint;
-
-        public event EventHandler DelayedTextChanged;
-        private EventHandler _hintActiveChanged;
-        /// <summary>
-        /// will be triggered when IsHintActive changes.
-        /// this event will detect when a callback is already registered and will not register again
-        /// </summary>
-        public event EventHandler HintActiveChanged
+        add
         {
-            add
+            if (_hintActiveChanged == null || !_hintActiveChanged.GetInvocationList().Contains(value))
             {
-                if (_hintActiveChanged == null || !_hintActiveChanged.GetInvocationList().Contains(value))
-                {
-                    _hintActiveChanged += value;
-                }
-            }
-            remove
-            {
-                _hintActiveChanged -= value;
+                _hintActiveChanged += value;
             }
         }
-
-        private Color borderColor = Color.Blue;
-        public Color BorderColor
+        remove
         {
-            get => borderColor;
-            set
-            {
-                borderColor = value;
-            }
+            _hintActiveChanged -= value;
         }
-        [Editor]
-        public int DelayedTextChangedTimeout { get; set; }
+    }
 
-        [Editor]
-        [Description("If set, this text will be shown as grayed hint until the user enters the box")]
-        public string Hint
+    private Color borderColor = Color.Blue;
+    public Color BorderColor
+    {
+        get => borderColor;
+        set
         {
-            get => _hint;
-            set
-            {
-                _hint = value;
-                if (string.IsNullOrEmpty(base.Text))
-                    base.Text = value;
-            }
+            borderColor = value;
         }
+    }
+    [Editor]
+    public int DelayedTextChangedTimeout { get; set; }
 
-        public bool IsHintActive { get; private set; }
-
-        public bool IsDelayActive { get; set; } = true;
-
-        public new string Text => base.Text == Hint ? "" : base.Text;
-
-        public StylableTextBox()
+    [Editor]
+    [Description("If set, this text will be shown as grayed hint until the user enters the box")]
+    public string Hint
+    {
+        get => _hint;
+        set
         {
-            DelayedTextChangedTimeout = 900; // 0.9 seconds
-            IsHintActive = true;
-            BorderStyle = BorderStyle.None;
+            _hint = value;
+            if (string.IsNullOrEmpty(base.Text))
+                base.Text = value;
         }
+    }
 
-        protected override void Dispose(bool disposing)
+    public bool IsHintActive { get; private set; }
+
+    public bool IsDelayActive { get; set; } = true;
+
+    public new string Text => base.Text == Hint ? "" : base.Text;
+
+    public StylableTextBox()
+    {
+        DelayedTextChangedTimeout = 900; // 0.9 seconds
+        IsHintActive = true;
+        BorderStyle = BorderStyle.None;
+    }
+
+    protected override void Dispose(bool disposing)
+    {
+        if (m_delayedTextChangedTimer != null)
         {
-            if (m_delayedTextChangedTimer != null)
-            {
-                m_delayedTextChangedTimer.Stop();
-                if (disposing)
-                    m_delayedTextChangedTimer.Dispose();
-            }
-
-            base.Dispose(disposing);
+            m_delayedTextChangedTimer.Stop();
+            if (disposing)
+                m_delayedTextChangedTimer.Dispose();
         }
+
+        base.Dispose(disposing);
+    }
         
 
-        #region timer events / methods
-        protected virtual void OnDelayedTextChanged(EventArgs e)
-        {
-            DelayedTextChanged?.Invoke(this, e);
-        }
-
-        protected override void OnTextChanged(EventArgs e)
-        {
-            if (IsDelayActive && DelayedTextChangedTimeout > 0)
-            {
-                InitializeDelayedTextChangedEvent();
-            }
-            base.OnTextChanged(e);
-        }
-
-        private void InitializeDelayedTextChangedEvent()
-        {
-            m_delayedTextChangedTimer?.Stop();
-
-            if (m_delayedTextChangedTimer == null || m_delayedTextChangedTimer.Interval != DelayedTextChangedTimeout)
-            {
-                m_delayedTextChangedTimer = new Timer();
-                m_delayedTextChangedTimer.Tick += HandleDelayedTextChangedTimerTick;
-                m_delayedTextChangedTimer.Interval = DelayedTextChangedTimeout;
-            }
-
-            m_delayedTextChangedTimer.Start();
-        }
-
-        private void HandleDelayedTextChangedTimerTick(object sender, EventArgs e)
-        {
-            if (sender is Timer timer)
-            {
-                timer.Stop();
-            }
-
-            OnDelayedTextChanged(EventArgs.Empty);
-        }
-        #endregion
-        #region hint events / methods
-
-        protected override void OnLostFocus(EventArgs e)
-        {
-            if (!IsHintActive && string.IsNullOrEmpty(base.Text))
-            {
-                IsHintActive = true;
-                base.Text = Hint;
-                if(_hintActiveChanged != null)
-                    _hintActiveChanged(this, EventArgs.Empty);
-            }
-
-            base.OnLostFocus(e);
-        }
-
-        protected override void OnGotFocus(EventArgs e)
-        {
-            if (IsHintActive && !string.IsNullOrEmpty(base.Text))
-            {
-                IsHintActive = false;
-                base.Text = "";
-                if (_hintActiveChanged != null)
-                    _hintActiveChanged(this, EventArgs.Empty);
-            }
-
-            base.OnGotFocus(e);
-        }
-        #endregion
+    #region timer events / methods
+    protected virtual void OnDelayedTextChanged(EventArgs e)
+    {
+        DelayedTextChanged?.Invoke(this, e);
     }
+
+    protected override void OnTextChanged(EventArgs e)
+    {
+        if (IsDelayActive && DelayedTextChangedTimeout > 0)
+        {
+            InitializeDelayedTextChangedEvent();
+        }
+        base.OnTextChanged(e);
+    }
+
+    private void InitializeDelayedTextChangedEvent()
+    {
+        m_delayedTextChangedTimer?.Stop();
+
+        if (m_delayedTextChangedTimer == null || m_delayedTextChangedTimer.Interval != DelayedTextChangedTimeout)
+        {
+            m_delayedTextChangedTimer = new Timer();
+            m_delayedTextChangedTimer.Tick += HandleDelayedTextChangedTimerTick;
+            m_delayedTextChangedTimer.Interval = DelayedTextChangedTimeout;
+        }
+
+        m_delayedTextChangedTimer.Start();
+    }
+
+    private void HandleDelayedTextChangedTimerTick(object sender, EventArgs e)
+    {
+        if (sender is Timer timer)
+        {
+            timer.Stop();
+        }
+
+        OnDelayedTextChanged(EventArgs.Empty);
+    }
+    #endregion
+    #region hint events / methods
+
+    protected override void OnLostFocus(EventArgs e)
+    {
+        if (!IsHintActive && string.IsNullOrEmpty(base.Text))
+        {
+            IsHintActive = true;
+            base.Text = Hint;
+            if(_hintActiveChanged != null)
+                _hintActiveChanged(this, EventArgs.Empty);
+        }
+
+        base.OnLostFocus(e);
+    }
+
+    protected override void OnGotFocus(EventArgs e)
+    {
+        if (IsHintActive && !string.IsNullOrEmpty(base.Text))
+        {
+            IsHintActive = false;
+            base.Text = "";
+            if (_hintActiveChanged != null)
+                _hintActiveChanged(this, EventArgs.Empty);
+        }
+
+        base.OnGotFocus(e);
+    }
+    #endregion
 }
