+++ StylableWinFormsControls/StylableWinFormsControls/Controls/StylableToolStrip.cs
@@ -1,26 +1,27 @@
-﻿namespace AssortedDevelopment.StylableWinFormsControls
+﻿using StylableWinFormsControls.Native;
+
+namespace StylableWinFormsControls;
+
+/// <summary>
+/// This class adds on to the functionality provided in System.Windows.Forms.ToolStrip.
+/// </summary>
+public class StylableToolStrip : ToolStrip
 {
+
     /// <summary>
-    /// This class adds on to the functionality provided in System.Windows.Forms.ToolStrip.
+    /// Gets or sets whether the ToolStripEx honors item clicks when its containing form does
+    /// not have input focus.
     /// </summary>
-    public class StylableToolStrip : ToolStrip
+    /// <remarks>
+    /// Default value is false, which is the same behavior provided by the base ToolStrip class.
+    /// </remarks>
+    public bool ClickThrough { get; set; } = false;
+
+    protected override void WndProc(ref Message m)
     {
+        base.WndProc(ref m);
 
-        /// <summary>
-        /// Gets or sets whether the ToolStripEx honors item clicks when its containing form does
-        /// not have input focus.
-        /// </summary>
-        /// <remarks>
-        /// Default value is false, which is the same behavior provided by the base ToolStrip class.
-        /// </remarks>
-        public bool ClickThrough { get; set; } = false;
-
-        protected override void WndProc(ref Message m)
-        {
-            base.WndProc(ref m);
-
-            if (ClickThrough && m.Msg == NativeMethods.WM_MOUSEACTIVATE && m.Result == (IntPtr)NativeMethods.MA_ACTIVATEANDEAT)
-                m.Result = (IntPtr)NativeMethods.MA_ACTIVATE;
-        }
+        if (ClickThrough && m.Msg == NativeMethods.WM_MOUSEACTIVATE && m.Result == (IntPtr)NativeMethods.MA_ACTIVATEANDEAT)
+            m.Result = (IntPtr)NativeMethods.MA_ACTIVATE;
     }
 }
